:bug: CI build 失败修护
