+++ lib_common/src/main/res/layout/common_cyxbs_toast.xml
@@ -10,5 +10,5 @@
     android:paddingBottom="9dp"
     android:gravity="center"
     android:background="@drawable/common_shape_cyxbs_toast"
-    android:textColor="@color/common_discover_academic_online_colors"
+    android:textColor="@color/config_white_black"
     tools:text="自定义Toast" />
