+++ lib_common/src/main/res/values/attrs.xml
@@ -1,38 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:tools="http://schemas.android.com/tools" >
-    <attr name="cardCornerRadius" format="dimension" />
-    <attr name="contentPadding" format="dimension" />
-    <attr name="contentPaddingLeft" format="dimension" />
-    <attr name="contentPaddingTop" format="dimension" />
-    <attr name="contentPaddingRight" format="dimension" />
-    <attr name="contentPaddingBottom" format="dimension" />
-
-    <declare-styleable name="JCardViewPlus">
-        <attr name="cardBackgroundColor" format="color" />
-        <attr name="shadowColor" format="color" />
-        <attr name="shadowRadius" format="dimension" />
-        <attr name="shadowFlags">
-            <flag name="left" value="1" />
-            <flag name="top" value="2" />
-            <flag name="right" value="4" />
-            <flag name="bottom" value="8" />
-        </attr>
-
-        <attr name="cardCornerRadius" />
-        <attr name="cardCornerRadiusUpperLeft" format="dimension" />
-        <attr name="cardCornerRadiusUpperRight" format="dimension" />
-        <attr name="cardCornerRadiusLowerLeft" format="dimension" />
-        <attr name="cardCornerRadiusLowerRight" format="dimension" />
-
-        <attr name="contentPadding" />
-        <attr name="verticalContentPadding" format="dimension" />
-        <attr name="horizontalContentPadding" format="dimension" />
-        <attr name="contentPaddingLeft" />
-        <attr name="contentPaddingTop" />
-        <attr name="contentPaddingRight" />
-        <attr name="contentPaddingBottom" />
-    </declare-styleable>
-
     <declare-styleable name="RatioLayout">
         <attr name="ratio" format="float" />
     </declare-styleable>
