+++ lib_crash/build.gradle.kts
@@ -3,6 +3,7 @@ plugins {
 }
 
 useARouter()
+useAutoService()
 
 dependencies {
   implementation(projects.apiInit)
