--- module_discover/src/main/res/drawable-xhdpi/discover_shape_jwzx_feeds.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <corners
-        android:topLeftRadius="16dp"
-        android:topRightRadius="16dp" />
-    <solid android:color="@color/common_secondary_window_background" />
-</shape>
