+++ module_discover/src/main/res/drawable/discover_shape_jwzx_feeds.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <corners
+        android:topLeftRadius="16dp"
+        android:topRightRadius="16dp" />
+    <solid android:color="@color/config_secondary_window_background" />
+</shape>
