+++ module_discover/src/main/res/drawable/discover_shape_jwzx_title.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <corners
+        android:bottomLeftRadius="14dp"
+        android:topRightRadius="14dp" />
+    <solid android:color="@color/discover_academic_online_colors_background" />
+</shape>
