+++ module_discover/src/main/res/drawable/discover_shape_more_function_recycler_item.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <corners android:radius="8dp" />
-    <solid android:color="@color/common_curriculum_background_color" />
+    <solid android:color="@color/config_common_background_color" />
 </shape>
