+++ module_discover/src/main/res/drawable/discover_shape_more_function_recycler_item_chosen.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <corners android:radius="8dp" />
-    <solid android:color="@color/common_chosen_background" />
+    <solid android:color="@color/config_chosen_background" />
 </shape>
