+++ module_discover/src/main/res/layout/discover_item_more_function.xml
@@ -19,7 +19,7 @@
         android:layout_gravity="center_horizontal"
         android:layout_marginTop="6dp"
         android:alpha="0.6"
-        android:textColor="@color/common_discovery_function_item_color"
+        android:textColor="@color/discovery_function_item_color"
         android:textSize="11sp"
         tools:text="教师查询" />
