+++ module_discover/src/main/res/values-night/colors.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <!--    发现教务在线颜色-->
-    <color name="common_discover_academic_online_colors">#ffffff</color>
-    <color name="common_discover_academic_online_colors_background">#AFD2FB</color>
+    <color name="discover_academic_online_colors">#000000</color>
+    <color name="discover_academic_online_colors_background">#5A5A5A</color>
     <!--    发现菜单字体颜色-->
-    <color name="common_menu_font_color_found">#15315B</color>
+    <color name="discover_menu_font_color_found">#5A5A5A</color>
     <!--    发现菜单字体颜色-->
-    <color name="common_discovery_function_item_color">#15315B</color>
+    <color name="discovery_function_item_color">#8C8C8C</color>
 </resources>
