--- module_electricity/src/main/res/drawable-xhdpi/electricity_shape_round.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="@color/common_white_background" />
-    <corners android:radius="8dp" />
-</shape>
