+++ module_electricity/src/main/res/drawable/electricity_shape_round.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/config_white_black" />
+    <corners android:radius="8dp" />
+</shape>
