+++ module_emptyroom/src/main/res/layout/emptyroom_recycle_item_query_option.xml
@@ -12,7 +12,7 @@
         android:paddingTop="2dp"
         android:paddingRight="15dp"
         android:paddingBottom="2dp"
-        android:textColor="@color/common_level_two_font_color"
+        android:textColor="@color/config_level_two_font_color"
         android:textSize="14sp"
         android:maxLines="1"
         tools:text="第一周" />
