+++ module_main/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-  id("manager.library")
+  id("manager.libraryApp")
 }
 
 // 测试使用，设置 module_main 暂时不依赖的模块
@@ -14,7 +14,7 @@ dependencies {
         && it != project
         && it.name != "module_app" // module_app 依赖 module_main，而不是反向依赖
         && it.name != "lib_single" // lib_single 只跟单模块调试有关，单模块编译时单独依赖
-        && it.name != "lib_debug" // lib_debug 会在 BaseApplicationPlugin 插件中单独依赖
+        && it.name != "lib_debug" // lib_debug 单独依赖
   }.forEach {
     "api"(it)
   }
@@ -23,6 +23,8 @@ dependencies {
 useARouter()
 
 dependencies {
+  debugImplementation(projects.libDebug)
+
   implementation(libs.bundles.projectBase)
   implementation(libs.bundles.views)
   implementation(libs.bundles.network)
