+++ module_mine/src/main/java/com/mredrock/cyxbs/mine/page/mine/widget/SlideViewGroup.kt
@@ -3,7 +3,6 @@ package com.mredrock.cyxbs.mine.page.mine.widget
 import android.animation.ValueAnimator
 import android.content.Context
 import android.util.AttributeSet
-import android.util.Log
 import android.view.MotionEvent
 import android.view.View
 import android.view.animation.OvershootInterpolator
@@ -58,8 +57,6 @@ class SlideViewGroup @JvmOverloads constructor(
     override fun onNestedScrollAccepted(child: View, target: View, axes: Int, type: Int) {
         valueAnimator?.cancel()
         mNestedScrollingParentHelper.onNestedScrollAccepted(child, target, axes, type)
-        Log.d("ggg", "(SlideViewGroup.kt:60)-->> " +
-                "onNestedScrollAccepted")
     }
